requirements.txt
Django==2.0.1
django-filter==1.1.0
django-mptt==0.9.0
djangorestframework==3.7.7
xlrd==1.1.0
lxml==4.2.0
pyquery==1.4.0
